+++ packages/relay/src/lib/services/ethService/ethCommonService/index.ts
@@ -23,8 +23,11 @@ import {JsonRpcError, predefined} from "../../../errors/JsonRpcError";
 import {ICommonService} from "./ICommonService";
 import { Logger } from 'pino';
 import { ClientCache, MirrorNodeClient } from '../../../clients';
-import {numberTo0x, parseNumericEnvVar} from "../../../../formatters";
+import { nullableNumberTo0x, numberTo0x, parseNumericEnvVar, toHash32 } from "../../../../formatters";
 import {SDKClientError} from "../../../errors/SDKClientError";
+import { MirrorNodeClientError } from '../../../errors/MirrorNodeClientError';
+import { Log } from '../../../model';
+import * as _ from 'lodash';
 
 
 /**
@@ -203,5 +206,103 @@ export class CommonService implements ICommonService {
             throw error;
         }
         return predefined.INTERNAL_ERROR(error.message.toString());
+    }
+
+    public async validateBlockHashAndAddTimestampToParams(params: any, blockHash: string, requestIdPrefix?: string) {
+        try {
+            const block = await this.mirrorNodeClient.getBlock(blockHash, requestIdPrefix);
+            if (block) {
+                params.timestamp = [
+                    `gte:${block.timestamp.from}`,
+                    `lte:${block.timestamp.to}`
+                ];
+            } else {
+                return false;
+            }
+        }
+        catch(e: any) {
+            if (e instanceof MirrorNodeClientError && e.isNotFound()) {
+                return false;
+            }
+
+            throw e;
+        }
+
+        return true;
+    }
+
+    public addTopicsToParams(params: any, topics: any[] | null) {
+        if (topics) {
+            for (let i = 0; i < topics.length; i++) {
+                if (!_.isNil(topics[i])) {
+                    params[`topic${i}`] = topics[i];
+                }
+            }
+        }
+    }
+
+    public async getLogsByAddress(address: string | [string], params: any, requestIdPrefix) {
+        const addresses = Array.isArray(address) ? address : [address];
+        const logPromises = addresses.map(addr => this.mirrorNodeClient.getContractResultsLogsByAddress(addr, params, undefined, requestIdPrefix));
+
+        const logResults = await Promise.all(logPromises);
+        const logs = logResults.flatMap(logResult => logResult ? logResult : [] );
+        logs.sort((a: any, b: any) => {
+            return a.timestamp >= b.timestamp ? 1 : -1;
+        });
+
+        return logs;
+    }
+
+    public async getLogsWithParams(address: string | [string] | null, params, requestIdPrefix?: string): Promise<Log[]> {
+        const EMPTY_RESPONSE = [];
+
+        let logResults;
+        if (address) {
+            logResults = await this.getLogsByAddress(address, params, requestIdPrefix);
+        }
+        else {
+            logResults = await this.mirrorNodeClient.getContractResultsLogs(params, undefined, requestIdPrefix);
+        }
+
+        if (!logResults) {
+            return EMPTY_RESPONSE;
+        }
+
+        const logs: Log[] = [];
+        for(const log of logResults) {
+            logs.push(
+              new Log({
+                  address: log.address,
+                  blockHash: toHash32(log.block_hash),
+                  blockNumber: numberTo0x(log.block_number),
+                  data: log.data,
+                  logIndex: nullableNumberTo0x(log.index),
+                  removed: false,
+                  topics: log.topics,
+                  transactionHash: toHash32(log.transaction_hash),
+                  transactionIndex: nullableNumberTo0x(log.transaction_index)
+              })
+            );
+        }
+
+        return logs;
+    }
+
+    public async getLogs(blockHash: string | null, fromBlock: string | 'latest', toBlock: string | 'latest', address: string | [string] | null, topics: any[] | null, requestIdPrefix?: string): Promise<Log[]> {
+        const EMPTY_RESPONSE = [];
+        const params: any = {};
+
+        if (blockHash) {
+            if ( !(await this.validateBlockHashAndAddTimestampToParams(params, blockHash, requestIdPrefix)) ) {
+                return EMPTY_RESPONSE;
+            }
+        } else if ( !(await this.validateBlockRangeAndAddTimestampToParams(params, fromBlock, toBlock, requestIdPrefix)) ) {
+            return EMPTY_RESPONSE;
+        }
+
+        this.addTopicsToParams(params, topics);
+
+        return this.getLogsWithParams(address, params, requestIdPrefix);
     }
 }
