+++ packages/server/src/koaJsonRpc/lib/methodConfiguration.ts
@@ -161,5 +161,8 @@ export const methodConfiguration = {
   },
   eth_uninstallFilter: {
     total: tier2rateLimit,
+  },
+  eth_getFilterLogs: {
+    total: tier2rateLimit,
   }
 };
